# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Application API document",
        "version": "1.0.0"
    },
    "paths": {
        "/api/registerAuth": {
            "post": {
                "summary": "Register User",
                "operationId": "e33f1826a6c48543523951b30637b68d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "false"
                    }
                }
            }
        },
        "/api/updateUser/{id}": {
            "put": {
                "summary": "Update User",
                "operationId": "a304952143f474e9673b8f6473055723",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "The position of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "description": "The department of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "description": "The age of the user",
                        "schema": {
                            "type": "date-time"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "The location of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project",
                        "in": "query",
                        "description": "The project of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sex",
                        "in": "query",
                        "description": "The sex of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permanent_address",
                        "in": "query",
                        "description": "The permanent_address of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "seniority",
                        "in": "query",
                        "description": "The seniority of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contract",
                        "in": "query",
                        "description": "The contract of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "temporary_address",
                        "in": "query",
                        "description": "The temporary_address of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "issued_by",
                        "in": "query",
                        "description": "The issued_by of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "personal_email",
                        "in": "query",
                        "description": "The personal_email of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "CCCD",
                        "in": "query",
                        "description": "The CCCD of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tax_code",
                        "in": "query",
                        "description": "The tax_code of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "leave_days",
                        "in": "query",
                        "description": "The leave_days of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "use_property",
                        "in": "query",
                        "description": "The use_property of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "The avatar of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "working_day",
                        "in": "query",
                        "description": "The working_day of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "promotion_day",
                        "in": "query",
                        "description": "The promotion_day of the user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_range",
                        "in": "query",
                        "description": "The date_range of the user",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "false"
                    }
                }
            }
        },
        "/api/getUser/{id}": {
            "get": {
                "summary": "Get User",
                "operationId": "970d55d515b1aa466e74ccee7ad31827",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "message: delete user successfull,user delete: {id: 7, email: a@123,  name: 123,...}"
                    },
                    "404": {
                        "description": "status:false,message: Call to a member function first() on null"
                    }
                }
            }
        },
        "/api/getUserLoggin": {
            "get": {
                "summary": "Get current logged-in user",
                "description": "Retrieve the information of the current logged-in user",
                "operationId": "0cbfa7b60ccee240cb651256c0db6703",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/deleteUser/{id}": {
            "delete": {
                "summary": "Delete User",
                "operationId": "3cedaf2c8991b02a8ce1073a1a0fb020",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK, message:delete user successfull,data: detail user,example : user: id=1,name=abc,..."
                    },
                    "404": {
                        "description": "status:false,message: Call to a member function first() on null"
                    }
                }
            }
        },
        "/api/loggin": {
            "post": {
                "summary": "Loggin User",
                "operationId": "97f37ed41999efa01a348b1689f7176f",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "message: loggin successfull,status: true"
                    },
                    "500": {
                        "description": "status:false, message: email not isset,"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "summary": "Logout User",
                "operationId": "b5b0e4026ed3cc40a861d3e9c4b7f0e4",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status:true,  message: logout successfull"
                    },
                    "500": {
                        "description": "status:false"
                    }
                }
            }
        },
        "/api/changePassword": {
            "post": {
                "summary": "changePassword User",
                "operationId": "151872945dc8b73689bf8e6ae53cd202",
                "parameters": [
                    {
                        "parameter": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "old_password",
                        "in": "query",
                        "description": "The old_password of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "description": "The new_password of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "message: Success'"
                    },
                    "422": {
                        "description": "status:false, message: Mật khẩu cũ không chính xác,"
                    },
                    "500": {
                        "description": "status:false"
                    }
                }
            }
        }
    }
}